requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
jinja2
pytest
pytest-asyncio